fix header title and icon using flex instead of line-height used earlier
